+++ FIASUpdate/Forms/FormAddressSearch.cs
@@ -9,7 +9,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace FIASUpdate
+namespace FIASUpdate.Forms
 {
     public partial class FormAddressSearch : Form
     {
