+++ FIASUpdate/Readers/FIASReader.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace FIASUpdate
+namespace FIASUpdate.Readers
 {
     internal class FIASReader : XMLDataReader
     {
